requirements.txt
google.generativeai
PyPDF2
python-dotenv
streamlit
streamlit-option-menu
